Share catalog routing context across chat surfaces
Move catalog-question detection into the routing package so raw chat route and wrapper interact surfaces share the same OMH catalog intuition. Extend natural coding delegation and memory-curation triggers so Hermes can route Codex implementation handoff and memory/context cleanup requests without falling back to generic clarification.

Constraint: Hermes-facing chat routing needs stronger OMH context without adding LLM, network, transport, or hidden execution behavior.

Rejected: Keep catalog detection only in wrapper contract | raw chat route would still miss the same natural workflow questions that wrappers can answer.

Confidence: high

Scope-risk: moderate

Directive: Keep install/verify command questions out of the workflow picker; route catalog discovery only when the user asks for available OMH workflows/skills/commands.

Tested: PYTHONPATH=tests uv run python -m unittest discover -s tests -v; uv run python -m src.cli docs workflows --check; uv run python -m src.cli release skill-content-smoke --json; uv run python -m compileall -q src tests && git diff --check; uv build.

Not-tested: Live Hermes/Discord rendering was not observed.
