+++ src/routing/policy.py
@@ -400,8 +400,16 @@ _CODING_HANDOFF_EXECUTOR_TOKENS = _normalized_token_set(
         "omo",
         "omc",
         "executor",
+        "codex로",
+        "codex에게",
+        "claude로",
+        "claude에게",
         "코덱스",
+        "코덱스로",
+        "코덱스에게",
         "클로드",
+        "클로드로",
+        "클로드에게",
     }
 )
 _CODING_HANDOFF_WORK_TOKENS = _normalized_token_set(
@@ -413,8 +421,10 @@ _CODING_HANDOFF_WORK_TOKENS = _normalized_token_set(
         "fix",
         "issue",
         "pr",
+        "feature",
         "구현",
         "코딩",
+        "기능",
         "수정",
         "고쳐",
         "이슈",
@@ -434,7 +444,12 @@ _CODING_HANDOFF_CONTROL_TOKENS = _normalized_token_set(
         "attach",
         "맡기고",
         "맡기",
+        "맡겨",
+        "맡겨줘",
+        "맡겨주세요",
         "넘기",
+        "넘겨",
+        "넘겨줘",
         "위임",
         "추적",
         "진행상태",
@@ -451,6 +466,9 @@ _CODING_HANDOFF_PHRASES = (
     "codex handoff",
     "codex progress tracking",
     "codex session tracking",
+    "codex로 이 기능 구현",
+    "codex로 구현 맡겨",
+    "codex로 맡겨",
     "track coding progress",
     "coding agent progress",
     "open in codex",
