+++ src/skills/catalog.py
@@ -2308,14 +2308,24 @@ _FEATURE_SURFACE_SKILLS = (
             "memory curation",
             "memory review",
             "memory inspect",
+            "memory context review",
+            "context review",
+            "context cleanup",
+            "context curation",
             "curate memory",
             "stale memory",
             "conflicting memory",
             "duplicate skill",
             "MEMORY.md",
             "USER.md",
+            "기억 점검",
             "기억 정리",
+            "기억하는 맥락",
             "메모리 정리",
+            "맥락 점검",
+            "맥락 정리",
+            "등록된 맥락",
+            "헤르메스 기억",
             "중복 스킬",
         ),
         "Use when Hermes memory, USER/MEMORY files, or accumulated skill guidance needs human-approved cleanup.",
